enable rocMLIR on Windows

+++ mlir/test/CAPI/mixr_cobj.cpp
@@ -206,10 +206,10 @@ static bool constructAndTraverseIr(MlirContext ctx) {
     for (auto arg : args) {
       argIdx += 3; // 3 per memref : allocated ptr, aligned ptr, offset
       auto sType = arg.getType().template cast<mlir::ShapedType>();
-      int64_t rank = sType.getRank();
-      printf("rank:%ld, dim:", rank);
+      long long rank = sType.getRank();
+      printf("rank:%lld, dim:", rank);
       for (int64_t i = 0; i < rank; ++i)
-        printf("<%ld>", sType.getDimSize(i));
+        printf("<%lld>", static_cast<long long>(sType.getDimSize(i)));
       printf("\n");
       argIdx += rank * 2; // 2 per each dimension : size, stride
     }
